backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.3
numpy==1.26.2
PyYAML==6.0.1
openpyxl==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.5
reportlab==4.0.7
openpyxl==3.1.2


